+++ tests/core.test.ts
@@ -1,7 +1,19 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import { rational } from "@";
 
+describe("Rational constructor", () => {
+	it("Accepts a class instance input", () => {
+		const instance = rational(1, 2);
+		expect(rational(instance).toString()).toBe("1/2");
+	});
+});
+
 describe("Parsing", () => {
+	it("Parses an integer input", () => {
+		expect(rational().toString()).toBe("0/1");
+		expect(rational(5).toString()).toBe("5/1");
+		expect(rational(-2).toString()).toBe("-2/1");
+	});
 	it("Parses two integer input", () => {
 		expect(rational(5, 8).toString()).toBe("5/8");
 		expect(rational(25, 8).toString()).toBe("25/8");
@@ -12,19 +24,40 @@ describe("Parsing", () => {
 		expect(rational(0.05).toString()).toBe("1/20");
 		expect(rational(3.2558).toString()).toBe("16279/5000");
 	});
-	it("Parses tuple input", () => {
+	it("Parses ratio input", () => {
+		expect(rational([]).toString()).toBe("0/1");
 		expect(rational([ 5, 8 ]).toString()).toBe("5/8");
 		expect(rational([ 25, 8 ]).toString()).toBe("25/8");
 		expect(rational([ 89, 2 ]).toString()).toBe("89/2");
 		expect(rational([ 5 ]).toString()).toBe("5/1");
 	});
 	it("Parses fraction object input", () => {
+		expect(rational({ n: 5 }).toString()).toBe("5/1");
+		expect(rational({ n: -2 }).toString()).toBe("-2/1");
 		expect(rational({ n: 5, d: 8 }).toString()).toBe("5/8");
+		expect(rational({ n: -5, d: 8 }).toString()).toBe("-5/8");
+		expect(rational({ n: 5, d: -8 }).toString()).toBe("-5/8");
 		expect(rational({ n: 25, d: 8 }).toString()).toBe("25/8");
 		expect(rational({ n: 89, d: 2 }).toString()).toBe("89/2");
 		expect(rational({ int: 1, n: 5, d: 8 }).toString()).toBe("13/8");
 		expect(rational({ int: -1, n: 7, d: 8 }).toString()).toBe("-15/8");
 		expect(rational({ int: 1, n: -89, d: 2 }).toString()).toBe("91/2");
+	});
+	it("Parses repeting decimal object", () => {
+		expect(rational({ repeat: "1" }).toString()).toBe("1/9");
+		expect(rational({ repeat: 2 }).toString()).toBe("2/9");
+		expect(rational({ nonrepeat: "0", repeat: "1" }).toString()).toBe("1/90");
+		expect(rational({ nonrepeat: 0, repeat: 1 }).toString()).toBe("1/90");
+		expect(rational({ nonrepeat: "0", repeat: 1 }).toString()).toBe("1/90");
+		expect(rational({ nonrepeat: "12", repeat: "45" }).toString()).toBe("137/1100");
+		expect(rational({ sign: -1, nonrepeat: "12", repeat: "45" }).toString()).toBe("-137/1100");
+		expect(rational({ sign: 1, int: 0, nonrepeat: "12", repeat: "45" }).toString()).toBe("137/1100");
+		expect(rational({ int: 10, nonrepeat: 1, repeat: 2 }).toString()).toBe("911/90");
+		expect(rational({ sign: -1, int: 10, nonrepeat: 1, repeat: 2 }).toString()).toBe("-911/90");
+		expect(rational({ int: 10, repeat: 2 }).toString()).toBe("92/9");
+		expect(rational({ sign: -1, int: 10, repeat: 2 }).toString()).toBe("-92/9");
+		expect(rational({ sign: -1, int: 10, repeat: "Hi!" }).toString()).toBe("0/1");
+		expect(rational({ sign: -1, nonrepeat: "gh", repeat: "Hi!" }).toString()).toBe("0/1");
 	});
 	it("Parses degrees object input", () => {
 		expect(rational({ deg: 1, min: 2, sec: 3 }).toString()).toBe("1241/1200");
@@ -33,7 +66,7 @@ describe("Parsing", () => {
 		expect(rational({ deg: 3 }).toString()).toBe("3/1");
 		expect(rational({ deg: 2, sec: 2 }).toString()).toBe("3601/1800");
 	});
-	it("Parses string with fractional form", () => {
+	it("Parses fractional string", () => {
 		expect(rational("1/2").toString()).toBe("1/2");
 		expect(rational("4/12").toString()).toBe("1/3");
 		expect(rational("27/9").toString()).toBe("3/1");
@@ -44,7 +77,7 @@ describe("Parsing", () => {
 		expect(rational("2 -1/2").toString()).toBe("5/2");
 		expect(rational("4 -4/-12").toString()).toBe("13/3");
 	});
-	it("Parses string with repeting decimal form", () => {
+	it("Parses repeting decimal string", () => {
 		expect(rational(".(1)").toString()).toBe("1/9");
 		expect(rational(".0(1)").toString()).toBe("1/90");
 		expect(rational(".12(45)").toString()).toBe("137/1100");
@@ -54,17 +87,7 @@ describe("Parsing", () => {
 		expect(rational("10.(2)").toString()).toBe("92/9");
 		expect(rational("-10.(2)").toString()).toBe("-92/9");
 	});
-	it("Parses repeting decimal object", () => {
-		expect(rational({ repeat: "1" }).toString()).toBe("1/9");
-		expect(rational({ nonrepeat: "0", repeat: "1" }).toString()).toBe("1/90");
-		expect(rational({ nonrepeat: "12", repeat: "45" }).toString()).toBe("137/1100");
-		expect(rational({ sign: -1, nonrepeat: "12", repeat: "45" }).toString()).toBe("-137/1100");
-		expect(rational({ int: 10, nonrepeat: 1, repeat: 2 }).toString()).toBe("911/90");
-		expect(rational({ sign: -1, int: 10, nonrepeat: 1, repeat: 2 }).toString()).toBe("-911/90");
-		expect(rational({ int: 10, repeat: 2 }).toString()).toBe("92/9");
-		expect(rational({ sign: -1, int: 10, repeat: 2 }).toString()).toBe("-92/9");
-	});
-	it("Parses string with degrees value", () => {
+	it("Parses degrees value string", () => {
 		expect(rational("1.12'5''").toString()).toBe("173/144");
 		expect(rational("-1.2'5''").toString()).toBe("-149/144");
 		expect(rational("7'5''").toString()).toBe("17/144");
@@ -82,7 +105,7 @@ describe("Parsing", () => {
 		expect(rational({ n: 25, d: -8 }).toString()).toBe("-25/8");
 		expect(rational({ n: -89, d: -2 }).toString()).toBe("89/2");
 	});
-	it("Tells about unsupported input", () => {
+	it("Provides information about unsupported input", () => {
 		expect(rational(2 ** 53 + 1).valid).toBe(false);
 		expect(rational("fggfg").valid).toBe(false);
 		expect(rational("25/0").valid).toBe(false);
@@ -91,6 +114,10 @@ describe("Parsing", () => {
 		expect(rational({ num: 25, den: 8 }).valid).toBe(false);
 		// @ts-expect-error
 		expect(rational([ 5, 2, 1 ]).valid).toBe(false);
+		// @ts-expect-error
+		expect(rational([ "str", "str" ]).valid).toBe(false);
+		// @ts-ignore
+		expect(rational({ n: "O", d: "3" }).valid).toBe(false);
 	});
 	it("Do not accept zero as denominator", () => {
 		expect(rational("56/0").valid).toBe(false);
@@ -134,13 +161,25 @@ describe("Representation", () => {
 });
 
 describe("Properties", () => {
-	it("Calculates the integral part", () => {
+	it("Returns the numerator", () => {
+		expect(rational(25, 40).numerator).toBe(5);
+		expect(rational([ 54, 12 ]).numerator).toBe(9);
+		expect(rational({ n: 25, d: 40 }).numerator).toBe(5);
+		expect(rational([ 11, 10 ]).numerator).toBe(11);
+	});
+	it("Returns the denominator", () => {
+		expect(rational(25, 40).denominator).toBe(8);
+		expect(rational([ 54, 12 ]).denominator).toBe(2);
+		expect(rational({ n: 25, d: 40 }).denominator).toBe(8);
+		expect(rational([ 11, 10 ]).denominator).toBe(10);
+	});
+	it("Returns the integral part", () => {
 		expect(rational(25, 40).integralPart).toBe(0);
 		expect(rational([ 54, 12 ]).integralPart).toBe(4);
 		expect(rational({ n: 25, d: 40 }).integralPart).toBe(0);
 		expect(rational([ 11, 10 ]).integralPart).toBe(1);
 	});
-	it("Calculates the fractional part", () => {
+	it("Returns the fractional part", () => {
 		expect(rational(25, 40).fractionalPart.toString()).toBe("5/8");
 		expect(rational([ 54, 12 ]).fractionalPart.toString()).toBe("1/2");
 		expect(rational({ n: 25, d: 40 }).fractionalPart.toString()).toBe("5/8");
@@ -152,26 +191,27 @@ describe("Properties", () => {
 		expect(rational({ n: 25, d: 40 }).proper).toBe(true);
 		expect(rational([ 11, 10 ]).proper).toBe(false);
 	});
-	it("Creates the reciprocal", () => {
+	it("Calculates the reciprocal", () => {
 		expect(rational(25, 40).reciprocal.toString()).toBe("8/5");
 		expect(rational([ 54, 12 ]).reciprocal.toString()).toBe("2/9");
 		expect(rational({ n: 25, d: 40 }).reciprocal.toString()).toBe("8/5");
 		expect(rational([ 11, 10 ]).reciprocal.toString()).toBe("10/11");
 	});
-	it("Returns the ratio's sign", () => {
+	it("Returns the sign", () => {
 		expect(rational(1, 2).sign).toBe(1);
 		expect(rational([ -5, 8 ]).sign).toBe(-1);
 		expect(rational({ n: 25, d: -8 }).sign).toBe(-1);
 		expect(rational({ n: -89, d: -2 }).sign).toBe(1);
 	});
-	it("Creates the opposite rational number", () => {
+	it("Calculates the opposite rational number", () => {
 		expect(rational(1, 2).opposite.sign).toBe(-1);
 		expect(rational([ -5, 8 ]).opposite.sign).toBe(1);
 		expect(rational({ n: 25, d: -8 }).opposite.sign).toBe(1);
 		expect(rational({ n: -89, d: -2 }).opposite.sign).toBe(-1);
 	});
-	it("Detects if the rational number can be represented as repeating decimal", () => {
+	it("Detects the repeating decimal", () => {
 		expect(rational(1, 2).repeating).toBe(false);
+		expect(rational(1, 25).repeating).toBe(false);
 		expect(rational(1, 9).repeating).toBe(true);
 		expect(rational(23, 90).repeating).toBe(true);
 		expect(rational(-156, 9990).repeating).toBe(true);
@@ -180,37 +220,37 @@ describe("Properties", () => {
 });
 
 describe("Operations", () => {
-	it("Adds two rational numbers", () => {
+	it("Performs summation", () => {
 		expect(rational(1, 2).add(rational({ n: 1, d: 4 })).toString()).toBe("3/4");
 		expect(rational({ n: 1, d: 2 }).add([ 1, 4 ]).toString()).toBe("3/4");
 		expect(rational(7, 8).add({ n: 5, d: 6 }).toString()).toBe("41/24");
 		expect(rational([ 7, 8 ]).add(5, 6).toString()).toBe("41/24");
 	});
-	it("Subtracts two rational numbers", () => {
+	it("Performs subtraction", () => {
 		expect(rational(1, 2).sub(rational({ n: 1, d: 4 })).toString()).toBe("1/4");
 		expect(rational({ n: 1, d: 2 }).sub([ 1, 4 ]).toString()).toBe("1/4");
 		expect(rational(7, 8).sub({ n: 5, d: 6 }).toString()).toBe("1/24");
 		expect(rational([ 7, 8 ]).sub(5, 6).toString()).toBe("1/24");
 	});
-	it("Multiplies two rational numbers", () => {
+	it("Performs multiplication", () => {
 		expect(rational(1, 2).mul(rational({ n: 1, d: 4 })).toString()).toBe("1/8");
 		expect(rational({ n: 1, d: 2 }).mul([ 1, 4 ]).toString()).toBe("1/8");
 		expect(rational(7, 8).mul({ n: 5, d: 6 }).toString()).toBe("35/48");
 		expect(rational([ 7, 8 ]).mul(5, 6).toString()).toBe("35/48");
 	});
-	it("Divides two rational numbers", () => {
+	it("Performs division", () => {
 		expect(rational(1, 2).div(rational({ n: 1, d: 4 })).toString()).toBe("2/1");
 		expect(rational({ n: 1, d: 2 }).div([ 1, 4 ]).toString()).toBe("2/1");
 		expect(rational(7, 8).div({ n: 5, d: 6 }).toString()).toBe("21/20");
 		expect(rational([ 7, 8 ]).div(5, 6).toString()).toBe("21/20");
 	});
-	it("Returns the absolute value of the rational number", () => {
+	it("Calculates the absolute value", () => {
 		expect(rational(1, 2).abs.sign).toBe(1);
 		expect(rational([ -5, 8 ]).abs.sign).toBe(1);
 		expect(rational({ n: 25, d: -8 }).abs.sign).toBe(1);
 		expect(rational({ n: -89, d: -2 }).abs.sign).toBe(1);
 	});
-	it("Compares the rational numbers", () => {
+	it("Performs the comparison", () => {
 		expect(rational(1, 2).compare(1, 2)).toBe(0);
 		expect(rational(1, 2).compare(8, 2)).toBe(-1);
 		expect(rational(1, 2).compare(1, 3)).toBe(1);
@@ -221,7 +261,7 @@ describe("Operations", () => {
 		expect(rational(-4.25).compare(10.11)).toBe(-1);
 		expect(rational(4.250001).compare(4.25)).toBe(1);
 	});
-	it("Rounds the value of rational number", () => {
+	it("Performs the rounding", () => {
 		expect(rational(1, 2).round()).toBe(1);
 		expect(rational(1, 3).round()).toBe(0);
 		expect(rational(3, 4).round()).toBe(1);
@@ -233,7 +273,7 @@ describe("Operations", () => {
 		expect(rational(23, 8).round(5)).toBe(2.875);
 		expect(rational(43, 9).round(7)).toBe(4.7777778);
 	});
-	it("Ceils the value of rational number", () => {
+	it("Performs the ceiling", () => {
 		expect(rational(1, 2).ceil()).toBe(1);
 		expect(rational(1, 3).ceil()).toBe(1);
 		expect(rational(3, 4).ceil()).toBe(1);
@@ -245,7 +285,7 @@ describe("Operations", () => {
 		expect(rational(29, 7).ceil(5)).toBe(4.14286);
 		expect(rational(43, 9).ceil(7)).toBe(4.7777778);
 	});
-	it("Ceils the value of rational number", () => {
+	it("Performs the flooring", () => {
 		expect(rational(1, 2).floor()).toBe(0);
 		expect(rational(1, 3).floor()).toBe(0);
 		expect(rational(3, 4).floor()).toBe(0);
@@ -257,15 +297,18 @@ describe("Operations", () => {
 		expect(rational(29, 7).floor(5)).toBe(4.14285);
 		expect(rational(43, 9).floor(7)).toBe(4.7777777);
 	});
-	it("Calculates the modulo of two rational numbers", () => {
+	it("Calculates the modulo", () => {
 		expect(rational("13/3").mod("7/8").toString()).toBe("5/6");
 		expect(rational("13/7").mod("19/11").toString()).toBe("10/77");
 	});
-	it("Calculates the mathematical modulo of two rational numbers", () => {
+	it("Calculates the mathematical modulo", () => {
 		expect(rational("-13/3").mathmod("7/8").toString()).toBe("1/24");
 		expect(rational("-13/7").mathmod("19/11").toString()).toBe("123/77");
 	});
-	it("Calculates the power of two rational numbers", () => {
+	it("Performs the exponentiation", () => {
+		expect(rational(2).pow(-2)?.toString()).toBe("1/4");
+		expect(rational(1, 2).pow(2)?.toString()).toBe("1/4");
+		expect(rational(1, 2).pow(-2)?.toString()).toBe("4/1");
 		expect(rational(2).pow("1/2")).toBeNull();
 		expect(rational(0).pow("1/2")?.toString()).toBe("0/1");
 		expect(rational(27).pow("2/3")?.toString()).toBe("9/1");
@@ -274,15 +317,19 @@ describe("Operations", () => {
 		expect(rational(1).pow("123/489")?.toString()).toBe("1/1");
 		expect(rational(4, 7).pow(15, 26)).toBeNull();
 		expect(rational(64, 49).pow(4, 8)?.toString()).toBe("8/7");
+		expect(rational(-1, 2).pow(1, 2)).toBeNull();
+		expect(rational(-0.2).pow(0.5)).toBeNull();
+		expect(rational(1, 9).pow(1, 2)?.toString()).toBe("1/3");
+		expect(rational(1, 9).pow(-1, 2)?.toString()).toBe("3/1");
 	});
-	it("Calculates the GCD of two rational numbers", () => {
+	it("Calculates the GCD", () => {
 		expect(rational(5, 8).gcd(3, 7).toString()).toBe("1/56");
 		expect(rational(2, 3).gcd(7, 5).toString()).toBe("1/15");
 	});
-	it("Calculates the LCM of two rational numbers", () => {
+	it("Calculates the LCM", () => {
 		expect(rational(5, 8).lcm(3, 7).toString()).toBe("15/1");
 	});
-	it("Checks divisibility of two rational numbers", () => {
+	it("Performs the divisibility check", () => {
 		expect(rational(20, 8).divisible(1, 4)).toBe(true);
 		expect(rational(-20, -8).divisible(1, -4)).toBe(true);
 		expect(rational(-20, 8).divisible(-1, 4)).toBe(true);