+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Rational
 
+## 0.5.1 (2021-08-22)
+
+- [improvement]: Improved test coverage to 99.63%;
+- [fix]: Invalid results for exponentiation;
+- [fix]: Invalid type handling for repeating decimal inputs;
+
 ## 0.5.0 (2021-08-19)
 
 - [feature]: new type of input, degrees value object;